infra/signal/requirements.txt
aiohttp==3.6.2
python-socketio>=5.0.0
